pyproject.toml
[project]
name = "intelliapply"
version = "0.1.0"
description = "Intelligent job application tracking powered by AI - LLM-powered information extraction, smart search, and multi-status workflow management"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Jason Liao", email = "jason@example.com"}
]
keywords = ["job", "application", "tracker", "ai", "llm", "career"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]

dependencies = [
    "openpyxl>=3.0.0",
    "pandas>=1.3.0",
    "openai>=1.0.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "pydantic>=2.0.0",
    "playwright>=1.40.0",
    "pyyaml>=6.0.0"
]

[project.urls]
Homepage = "https://github.com/Jason-LJQ/IntelliApply"
Documentation = "https://github.com/Jason-LJQ/IntelliApply/blob/master/README.md"
Repository = "https://github.com/Jason-LJQ/IntelliApply"
Issues = "https://github.com/Jason-LJQ/IntelliApply/issues"

[project.scripts]
intelliapply = "intelliapply.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["intelliapply", "intelliapply.config", "intelliapply.utils"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]
"intelliapply.config" = ["credential-example.yaml"]
